requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
tqdm>=4.66
requests>=2.31
tenacity>=8.2
python-dotenv>=1.0
